+++ lua/CompMod/Commander/Alien/SupplyChanges/Post/BalanceMisc.lua
@@ -1,0 +1,5 @@
+kDrifterSupply = 5
+kWhipSupply = 30
+kCragSupply = 25
+kShadeSupply = 25
+kShiftSupply = 25