+++ lua/CompMod/Commander/Marine/SupplyChanges/Post/BalanceMisc.lua
@@ -1,0 +1,8 @@
+kMACSupply = 5
+kArmorySupply = 0
+kARCSupply = 25
+kRoboticsFactorySupply = 0
+
+-- Additional supply cost values
+kSentryBatterySupply = 10
+kObservatorySupply = 40